Improve the admin display.
This adds download links in debug mode.  For production mode we will
need to change how that is done.

The order of the columns in the various display could be made more
consistent.  It may be slow when we have a large database to show in
results the name of the algorithm / scorer and dataset / groundtruth
that were used to produce the result.