Add modules to better manage files.
Without this, deleting a model from Django or updating the file
associated with a model does not delete that file from storage.

From a provenance standpoint we may want to not allow deleting or
changing any file that has been used (for instance, if an algorithm has
been run at all, then changing it would actually mark the current
version as inactive and create a new version rather than modifying the
existing version).