Validate dockers.
This asynchronously validates that the algorithm and score algorithm are
dockers and extracts the docker image attributes.  A task is
automatically run on the worker whenever the algorithm or score
algorithm is saved.  This ensures that the uploaded tar file is actually
a docker tar file with a single image.