Run docker with --gpus flag.
If there are available GPUs, this adds the --gpus flag to the docker run
command.

Like in girder worker, we could check if we think it needs GPUs based on
the docker image labels having nvidia information.  Otherwise, we need
some way to know if we need a GPU worker, and, if so, how many GPUs are
required.