This allows using a docker-based worker.
Some of our tasks run docker commands.  To do this within a docker
container, it is necessary to mount the docker command and the docker
socket.  This uses the default locations of each of these as deployed in
ubuntu.

Further, one of the virtues of developing with docker is network
separation.  However, since the ports were specified in the
docker-compose.yml file, this wasn't full possible -- the docker
development environment has no need to expose these to the host
computer, and if the host already had one these ports in use it would
fail.  This moves the port exposure to
`docker-compose.override.native.yml`, and the instructions for using the
native python have changed appropriately.