Allow downloading files from dockerized minio.
The download url for dockerized minio includes the docker internal
network name and port (e.g., starting with `http://minio:9000/`).  DNS
should not be expected to resolve this, not should forwarding that port
be needed for development.

This adds an api endpoint
(`/api/download/<snake_case_model>/<id>/<field>`) which will download
any FieldFile record in a core model.  The download is provided with the
internal name of the file and, if known, a stored mimetype.

The download_file function is probably too open (though we should only
load models based on some user authorization).  The explicit check for
`//minio:` is ugly and brittle.