Fix running scoring algorithms in docker deployment.
When the worker is NOT a docker container, we can run a scoring
algorithm, but when it IS a docker container, it fails.

Specifically, a scoring algorithm has two inputs: the output of an
algorithm and the groundtruth.  We run a scoring algorithm via `docker
run`.  The algorithm output is passed on stdin, the groundtruth is
mounted as a file with the path `/groundtruth.dat`.  The scoring results
are on stdout and logs on stderr.

When we are storing data on minio, the algorithm results and groundtruth
are moved to local files via `_field_file_to_local_path`.  These are in
the `/tmp` directory.  However, when we run a docker from within a
docker, those `/tmp` paths are inside the worker docker container, and
the scorer docker container can't reach them, even with a volume mount.

The solution is to mount a specific directory to share files between
docker containers.  In this case, we mount /tmp/rgd_worker in the
worker.  This directory has the same name on the host, and is therefore
reachable by the scoring algorithm docker container.

Also, guard against failure to parse an overall score.