Implement an initial simple raster search and display.
This also fixes some coordinate transform issues.  This requires
upgrading to django 3.1 (which is currently in rc1, and is not
official).  Unfortunately, there is definitely code in
djangorestframework with will fail in django 3.1 until drf has a new
release.

Specifically, there is ambiguity in how django 3.0 handles GDAL 3
transforms.  Compounding the problem, rasterio is bundled with a GDAL
2.x library.  Django should be using our more complete GDAL 3 library,
but depending on the import order of rasterio and osgeo, rasterio may
not emit wkt coordinates in a consistent manner.