Improve swagger descriptions.
Some swagger clients show summaries or descriptions rather than the
endpoint as the primary menu item.  As such, make sure these are
populated.

We should always populate the `operation_summary` if we specify an
`operation_description` in `swagger_auto_schema`.

If we add model __doc__ strings, when we auto-generate views to generate
swagger API, the first line become the summary and the rest of the
docstring becomes the description.