Change when on_commit hook is registered.
Before, we were registering an on_commit hook for a task in the pre_save
signal.  This moves it to the post_save signal.  This fixes an issue
when in a synchronous testing environment -- the save would trigger the
on_commit, which would immediately be performed.  This changed the model
but did not update it in the view that was present in the main save
loop.  When the save was performed, any changes were lost.  By moving
this to post_save, the changes are maintained.  However, in a
factory_boy factory, after generation, the instance needs to be
refreshed from the database to pick up those changes.

This works in our asynchronous production environment, and makes testing
a little cleaner (and doesn't perform certain tasks twice).